// src/vs/platform/agentHost/node/codex/protocol/generated/Resource.ts
// AUTOGENERATED - do not edit.
// Source: `codex app-server generate-ts --experimental`
// Generated from @openai/codex 0.134.0, licensed Apache-2.0.
// Modified by VS Code to rewrite imports and apply repository formatting.
// See README.md in this directory for provenance and licensing notes.
// Regenerate: npm run codex:gen-protocol
// Source file: Resource.ts

import type { JsonValue } from "./serde_json/JsonValue.js";

/**
 * A known resource that the server is capable of reading.
 */
export type Resource = { annotations?: JsonValue, description?: string, mimeType?: string, name: string, size?: number, title?: string, uri: string, icons?: Array<JsonValue>, _meta?: JsonValue, };