// src/vs/platform/agentHost/node/codex/protocol/generated/ResourceTemplate.ts
// AUTOGENERATED - do not edit.
// Source: `codex app-server generate-ts --experimental`
// Generated from @openai/codex 0.134.0, licensed Apache-2.0.
// Modified by VS Code to rewrite imports and apply repository formatting.
// See README.md in this directory for provenance and licensing notes.
// Regenerate: npm run codex:gen-protocol
// Source file: ResourceTemplate.ts

import type { JsonValue } from "./serde_json/JsonValue.js";

/**
 * A template description for resources available on the server.
 */
export type ResourceTemplate = { annotations?: JsonValue, uriTemplate: string, name: string, title?: string, description?: string, mimeType?: string, };